+++ herovault.js
@@ -1,23 +1,72 @@
-let hvDebug = false;
-const hvVer="0.2.0";
+let hvDebug = true;
+const hvVer="0.3.0";
+const heroVaultURL='https://herovau.lt';
+// const heroVaultURL='https://dev.herovau.lt';
 
 const hvColor1='color: #7bf542';  //bright green
 const hvColor2='color: #d8eb34'; //yellow green
 const hvColor3='color: #ffffff'; //white
 const hvColor4='color: #cccccc'; //gray
 const hvColor5='color: #ff0000'; //red
-
-function isFunction(possibleFunction) {
-  return typeof(possibleFunction) === typeof(Function);
-}
+let HLOuserToken, hvUserToken,skipTokenPrompt;
+let enableHLO=true;
+let enablePB=true;
 
 Hooks.on('ready', async function() {
   console.log("%cHeroVau.lt/Foundry Bridge | %cinitializing",hvColor1,hvColor4);
+  if (Cookie.get('herovault_user_token')) {
+      hvUserToken=Cookie.get('herovault_user_token');
+      Cookie.set('herovault_user_token',hvUserToken,60);
+  }
+  if (Cookie.get('herovault_skiptoken')){
+      skipTokenPrompt=Cookie.get('herovault_skiptoken');
+      Cookie.set('herovault_skiptoken',"",-1);     
+  }
+      // game.settings.register('herovaultfoundry', 'userToken', {
+      //     name : "HeroVau.lt User Token",
+      //     hint : "Please enter your personal user token from "+heroVaultURL+". Your HeroVau.lt token allows you to import and export PCs directly into your HeroVau.lt account.  This is not required to use the Pathbuilder or HeroLab Online features.",
+      //     scope : 'world',
+      //     config : true,
+      //     type : String,
+      //     default : hvUserToken,
+      //     onChange: value =>  ( checkUserToken() )
+      // });
+      game.settings.register('herovaultfoundry', 'hlouserToken', {
+          name : "HeroLab Online User Token (optional)",
+          hint : "Please enter your personal user token. A user token allows external tools (like HeroVau.lt) to access the HLO server and perform export operations.",
+          scope : 'world',
+          config : true,
+          type : String,
+          default : '',
+          onChange: value =>  (setHLOToken())
+      });
+      game.settings.register('herovaultfoundry', 'debugEnabled', {
+          name : "Enable debug mode",
+          hint : "Debug output will be written to the js console.",
+          scope : 'world',
+          config : true,
+          type: Boolean,
+          default: false,
+          onChange: value => (hvDebug=game.settings.get('herovaultfoundry', 'debugEnabled'))
+      });
+      game.settings.register('herovaultfoundry', 'skipTokenPrompt', {
+          name : "Skip Token Prompt",
+          hint : "Once your HeroVau.lt user token is set, you will no longer be prompted to set it. Unchecking this makes HeroVau.lt prompt you for the User Token again.",
+          scope : 'world',
+          config : true,
+          type: Boolean,
+          default: false,
+          onChange: value => (skipTokenPrompt=game.settings.get('herovaultfoundry', 'skipTokenPrompt'))
+      });
+  hvDebug=game.settings.get('herovaultfoundry', 'debugEnabled');
+  HLOuserToken=game.settings.get('herovaultfoundry', 'hlouserToken');
+  hvUserToken=Cookie.get('herovault_user_token');
+  skipTokenPrompt=game.settings.get('herovaultfoundry', 'skipTokenPrompt');
 });
 
 Hooks.on('renderActorSheet', function(obj, html){
   const actor = obj.actor;
-  v8=versionCompare(game.data.version,'0.8.5');
+  const v8=HVversionCompare(game.data.version,'0.8.5');
   // Only inject the link if the actor is of type "character" and the user has permission to update it
   if (hvDebug) {
     if (v8==1)
@@ -26,7 +75,7 @@ Hooks.on('renderActorSheet', function(obj, html){
       console.log("%cHeroVau.lt/Foundry Bridge | %cActor type: " + actor.data.type + "can update?: " + actor.can(game.user, "update"),hvColor1,hvColor4);
   }
     
-  if (!actor.data.type === "character") return;
+  if (!(actor.data.type === "character")) return;
 
   if (v8==1) {
     if (actor.canUserModify(game.user, "update")==false) return;
@@ -40,10 +89,124 @@ Hooks.on('renderActorSheet', function(obj, html){
   
   let vaultButton = $(`<a class="popout"><i class="fas fa-cloud"></i>Vault</a>`);
 
-  vaultButton.on('click', () => beginVaultConnection(obj.object));
+  vaultButton.on('click', () => checkNextAction(obj.object));
   element.after(vaultButton);
+  if (game.modules.get('pathbuilder2e-import')?.active  && enablePB) {
+    $( "a:contains('Import from Pathbuilder')" ).remove();
+  }
 });
 
+function setHLOToken() {
+  HLOuserToken=Cookie.get('herovault_user_token');
+}
+
+function checkUserToken() {
+  var ut = Cookie.get('herovault_user_token');
+  var xmlhttp = new XMLHttpRequest();
+  xmlhttp.onreadystatechange = function() {
+    if (this.readyState == 4 && this.status == 200) {
+      let responseJSON = JSON.parse(this.responseText);
+      if (hvDebug) 
+          console.log("%cHeroVau.lt/Foundry Bridge | %c"+JSON.stringify(responseJSON),hvColor1,hvColor4);
+      if (responseJSON.status==1){
+        hvUserToken=ut;
+        game.settings.set('herovaultfoundry', 'skipTokenPrompt',true);
+        skipTokenPrompt=true;
+      } else {
+
+        hvUserToken="";
+        Cookie.set('herovault_user_token',"",-1);
+        game.settings.set('herovaultfoundry', 'skipTokenPrompt',false);    
+        skipTokenPrompt=false;
+      }
+    }
+  };
+  if (hvDebug)
+    console.log("%cHeroVau.lt/Foundry Bridge | %c/foundrymodule.php?action=iv&userToken="+ut,hvColor1,hvColor4);
+  xmlhttp.open("GET", heroVaultURL+"/foundrymodule.php?action=iv&userToken="+ut, true);
+  xmlhttp.send();
+}
+
+function checkNextAction(obj) {
+  if(!game.modules.get('herovaultfoundry')?.active) {
+    if (skipTokenPrompt) {
+      herovaultMenu(obj);
+    } else {
+      if (hvUserToken==null)
+        hvUserToken="";
+      getVaultToken(herovaultMenu,targetActor);
+    }
+  } else {
+    pickAFunction(obj);
+  }
+}
+
+async function loadPB(obj) {
+  let {beginPathbuilderImport} = await import('/modules/pathbuilder2e-import/pathbuilder-import.js');
+  beginPathbuilderImport(obj,true)
+}
+
+async function loadHLO(obj) {
+  import('../hlo-importer/hlo-importer.min.js').then(hlomodule => { let hlo = new hlomodule.HeroLabImporter; hlo.hloShim(obj); }).catch(err => {console.log("error importing hlo: "+ err.message)});
+}
+
+async function pickAFunction(obj) {
+  let hloImport = false;
+  let hvImport = false
+  let pbImport = false;
+  let dopt  = {
+    width: 400,
+    height: "auto"
+  }
+  let menuButtons ={
+    heroVaultImport: {
+      icon: "<i class='fas fa-cloud'></i>",
+      label: `HeroVau.lt Import/Export`,
+      callback: () => hvImport = true
+    }};
+  if (game.system.id=="pf2e" && game.modules.get('hlo-importer')?.active && enableHLO) {
+    menuButtons= {...menuButtons, hloimport: {
+          icon: "<i class='fas fa-flask'></i>",
+          label: `Import from Herolab Online`,
+          callback: () => hloImport = true
+    }};
+    dopt.width+=100;
+  }
+  if (game.system.id=="pf2e" && game.modules.get('pathbuilder2e-import')?.active && enablePB) {
+    menuButtons= {...menuButtons, pbimport: {
+          icon: "<i class='fas fa-check'></i>",
+          label: `Import from Pathbuilder 2e`,
+          callback: () => pbImport = true
+    }};
+    dopt.width+=100;
+  }
+  menuButtons= {...menuButtons, no: {
+    icon: "<i class='fas fa-times'></i>",
+    label: `Cancel`
+  }};
+
+  new Dialog({
+    title: `HeroVau.lt Import`,
+    content: `
+      <div>
+        <p>Please select the importer you'd like to use from the options below.</p>
+      <div>
+      <hr/>`,
+    buttons:menuButtons,
+    default: "no",
+    close: html => {
+      if (hvImport) {
+        beginVaultConnection(obj);
+      } else if (hloImport) {
+        loadHLO(obj)
+      } else if (pbImport) {
+        // import('../pathbuilder2e-import/pathbuilder-import.js').then(pbmodule => { beginPathbuilderImport(obj) }).catch(err => {console.log("error importing pb: "+ err.message)});
+        loadPB(obj);
+        // beginPathbuilderImport(obj);
+      }
+    }
+  }, dopt).render(true);   
+}
 /*
 Hooks.on('getSceneControlButtons', (controls) => {
     hvControls={
@@ -90,9 +253,8 @@ function importPFSChar() {}
 
 function getVaultToken(callback,callbackArg1,callbackArg2,callbackArg3,callbackArg4) {
   let applyChanges=false;
-  let defaulttoken=Cookie.get('herovault_user_token');
-  if (defaulttoken==null)
-    defaulttoken="";
+  if (hvUserToken==null)
+    hvUserToken="";
   new Dialog({
     title: `HeroVau.lt Import`,
     content: `
@@ -103,7 +265,7 @@ function getVaultToken(callback,callbackArg1,callbackArg2,callbackArg3,callbackA
       <div id="divCode">
         <div id="divOuter">
           <div id="divInner">
-            <input id="textBoxUserToken" type="text" maxlength="124" value="${defaulttoken}"/>
+            <input id="textBoxUserToken" type="text" maxlength="124" value="${hvUserToken}"/>
           </div>
         </div>
       </div>
@@ -154,9 +316,10 @@ function getVaultToken(callback,callbackArg1,callbackArg2,callbackArg3,callbackA
       if (applyChanges) {
         let userToken= html.find('[id="textBoxUserToken"]')[0].value;
         let skipToken= html.find('[id="skipToken"]')[0].checked;
-        Cookie.set('herovault_user_token',userToken,365); 
+        Cookie.set('herovault_user_token',userToken,60);
+        hvUserToken=userToken;
         if (skipToken)
-          Cookie.set('herovault_skiptoken',skipToken,30);
+          game.settings.set('herovaultfoundry', 'skipTokenPrompt',true);
         callback(callbackArg1,callbackArg2,callbackArg3,callbackArg4);
       }
     }
@@ -164,11 +327,14 @@ function getVaultToken(callback,callbackArg1,callbackArg2,callbackArg3,callbackA
 }
 
 
-async function exportToHV(targetActor,userToken) {
+async function exportToHV(targetActor) {
   try {
-    let exportNewPC = exportOverwritePC = vaultInfo = canOverwrite = false;
-    let hvUserToken=Cookie.get('herovault_user_token');
-    let portrait=hvUID="";
+    let exportNewPC = false;
+    let exportOverwritePC = false;
+    let vaultInfo = false;
+    let canOverwrite = false;
+    hvUserToken=Cookie.get('herovault_user_token');
+    let portrait,hvUID;
 
     if (targetActor.data.img.includes("mystery-man") == -1)
     {
@@ -185,7 +351,7 @@ async function exportToHV(targetActor,userToken) {
       canOverwrite=accChk;
       // Promise.resolve(checkForAccess(hvUserToken,hvUID)).then( res => canOverwrite=res);
     }
-    vaultInfo = await getVaultSlots(userToken);
+    vaultInfo = await getVaultSlots(hvUserToken);
     // Promise.resolve(getVaultSlots(userToken)).then( res => vaultInfo=res);
     if (hvDebug)
       console.log("%cHeroVau.lt/Foundry Bridge | %cvaultInfo: " + vaultInfo,hvColor1,hvColor4);
@@ -198,7 +364,7 @@ async function exportToHV(targetActor,userToken) {
     if (hvDebug)
       console.log("%cHeroVau.lt/Foundry Bridge | %ccan access?: " + canOverwrite,hvColor1,hvColor4);
     if (freeSlots <1 && canOverwrite==false){
-      bdy=`<div><p>Unfortunately you do not have enough open slots in your <a href="https://www.herovau.lt">HeroVau.lt</a> to import this PC.<br>Please upgrade your account or delete a PC from your account to free up some space.</p><div><hr/>`;
+      bdy=`<div><p>Unfortunately you do not have enough open slots in your <a href="https://herovau.lt">HeroVau.lt</a> to import this PC.<br>Please upgrade your account or delete a PC from your account to free up some space.</p><div><hr/>`;
 
       new Dialog({
         title: "Import to your HeroVau.lt",
@@ -244,7 +410,8 @@ async function exportToHV(targetActor,userToken) {
         default: "exportNew",
         close: async (html) => {
           if (exportNewPC) {
-            let exportStatus = await exportPCtoHV(targetActor, userToken, hvUID, true);
+            hvUID="";
+            let exportStatus = await exportPCtoHV(targetActor, hvUserToken, hvUID, true);
             if (exportStatus.error== true) {
               ui.notifications.error("Error exporting: " + exportStatus.message);
             } else {
@@ -253,7 +420,7 @@ async function exportToHV(targetActor,userToken) {
           } else if (exportOverwritePC){
             if (hvDebug)
               console.log("export overwrite PC");
-            let exportStatus = await exportPCtoHV(targetActor, userToken, hvUID, false);
+            let exportStatus = await exportPCtoHV(targetActor, hvUserToken, hvUID, false);
             if (exportStatus.error== true) {
               ui.notifications.error("Error exporting: " + exportStatus.message);
             } else {
@@ -286,9 +453,9 @@ const checkForAccess = (hvUserToken,hvUID) => {
       };
       if (hvDebug) {
         console.log("%cHeroVau.lt/Foundry Bridge | %cChecking if this account can access: " + hvUID,hvColor1,hvColor4);
-        console.log("%cHeroVau.lt/Foundry Bridge | %chttps://www.herovau.lt/foundrymodule.php?action=checkCharacter&userToken="+hvUserToken+"&charUID="+hvUID,hvColor1,hvColor4);
+        console.log("%cHeroVau.lt/Foundry Bridge | %chttps://herovau.lt/foundrymodule.php?action=checkCharacter&userToken="+hvUserToken+"&charUID="+hvUID,hvColor1,hvColor4);
       }
-      xmlhttp.open("GET", "https://www.herovau.lt/foundrymodule.php?action=checkCharacter&userToken="+hvUserToken+"&charUID="+hvUID, true);
+      xmlhttp.open("GET", heroVaultURL+"/foundrymodule.php?action=checkCharacter&userToken="+hvUserToken+"&charUID="+hvUID, true);
       xmlhttp.send();
     });
 };
@@ -310,8 +477,8 @@ const getVaultSlots = (hvUserToken) => {
       }
     };
     if (hvDebug)
-      console.log("%cHeroVau.lt/Foundry Bridge | %chttps://www.herovau.lt/foundrymodule.php?action=getVaultSlots&userToken="+hvUserToken,hvColor1,hvColor4);
-    xmlhttp.open("GET", "https://www.herovau.lt/foundrymodule.php?action=getVaultSlots&userToken="+hvUserToken, true);
+      console.log("%cHeroVau.lt/Foundry Bridge | %chttps://herovau.lt/foundrymodule.php?action=getVaultSlots&userToken="+hvUserToken,hvColor1,hvColor4);
+    xmlhttp.open("GET", heroVaultURL+"/foundrymodule.php?action=getVaultSlots&userToken="+hvUserToken, true);
     xmlhttp.send();
   });
 };
@@ -337,15 +504,21 @@ const exportPCtoHV = (targetActor, userToken,charUID,importAsNew) => {
         resolve(responseJSON);
       }
     };
-    // console.log("%cHeroVau.lt/Foundry Bridge | %chttps://www.herovau.lt/foundrymodule.php?action=importNewPC&userToken="+userToken+"&encodedChar="+pcEncodedJSON,hvColor1,hvColor4);
-    xmlhttp.open("POST", "https://www.herovau.lt/foundrymodule.php", true);
+    // console.log("%cHeroVau.lt/Foundry Bridge | %chttps://herovau.lt/foundrymodule.php?action=importNewPC&userToken="+userToken+"&encodedChar="+pcEncodedJSON,hvColor1,hvColor4);
+    xmlhttp.open("POST", heroVaultURL+"/foundrymodule.php", true);
     xmlhttp.setRequestHeader("Content-type", "application/x-www-form-urlencoded");
     xmlhttp.send('action='+action+'&userToken='+userToken+'&encodedChar='+pcEncodedJSON+'&gamesystem='+gameSystem+'&charUID='+charUID);
   });
 }
 
-function herovaultMenu(targetActor,userToken) {
-  let importPC = exportPC = PFSPC = hloPC = herolabEnabled = pfsEnabled = false;
+
+function herovaultMenu(targetActor) {
+  let importPC = false;
+  let exportPC = false;
+  let PFSPC = false;
+  let hloPC = false;
+  let herolabEnabled = false;
+  let pfsEnabled = false;
   let ttl=`HeroVau.lt Import`;
   let bdy=`<div><p>Please choose an action to perform:</p><div><hr/>`;
   let dopt  = {
@@ -377,7 +550,7 @@ function herovaultMenu(targetActor,userToken) {
           icon: "<i class='fas fa-flask'></i>",
           label: `Import from Herolab Online`,
           callback: () => hloPC = true
-        }};
+    }};
   }
   menuButtons= {...menuButtons, no: {
           icon: "<i class='fas fa-times'></i>",
@@ -390,11 +563,11 @@ function herovaultMenu(targetActor,userToken) {
     default: "yes",
     close: html => {
       if (importPC) {
-        loadPersonalVault(targetActor, userToken);
+        loadPersonalVault(targetActor, hvUserToken);
       } else if (exportPC){
         if (hvDebug)
           console.log("export PC");
-        exportToHV(targetActor,userToken);
+        exportToHV(targetActor,hvUserToken);
       } else if (PFSPC) {
         console.log("PFS PC");
       } else if (hloPC) {
@@ -404,28 +577,23 @@ function herovaultMenu(targetActor,userToken) {
   }, dopt).render(true); 
 }
 
-function exportPC(targetActor,userToken) {
+function exportPC(targetActor) {
   let applyChanges=false;
-  let defaulttoken=Cookie.get('herovault_user_token');
-  if (defaulttoken==null)
-    defaulttoken="";
+  if (hvUserToken==null)
+    hvUserToken="";
 }
 
-function beginVaultConnection(targetActor,userToken){
-  let applyChanges=false;
-  let defaulttoken=Cookie.get('herovault_user_token');
-  let skipTokenCookie=Cookie.get('herovault_skiptoken');
-  if (skipTokenCookie) {
-    herovaultMenu(targetActor, defaulttoken);
+function beginVaultConnection(targetActor){
+  if (skipTokenPrompt) {
+    herovaultMenu(targetActor, hvUserToken);
   } else {
-    if (defaulttoken==null)
-      defaulttoken="";
-    getVaultToken(herovaultMenu,targetActor, defaulttoken);
+    if (hvUserToken==null)
+      hvUserToken="";
+    getVaultToken(herovaultMenu,targetActor, hvUserToken);
   }
-
 }
 
-function loadPersonalVault(targetActor, userToken){
+function loadPersonalVault(targetActor){
     const gameSystem=game.data.system.id;
     let error=false;
     var xmlhttp = new XMLHttpRequest();
@@ -467,8 +635,9 @@ function loadPersonalVault(targetActor, userToken){
 
               createPCTable(targetActor, responseJSON);
           } else {
-            ui.notifications.warn("Unable load vault.  Please double-check your User Token.");
-            Cookie.set('herovault_skiptoken',"",-1);
+            ui.notifications.warn("Unable to load vault.  Please double-check your User Token.");
+            game.settings.set('herovaultfoundry', 'skipTokenPrompt',false);
+            getVaultToken(loadPersonalVault,targetActor, hvUserToken);
             return;
         }
       }
@@ -476,13 +645,14 @@ function loadPersonalVault(targetActor, userToken){
       // console.log("%cHeroVau.lt/Foundry Bridge | %creadyState: "+this.readyState,hvColor1,hvColor4)
       }
     };
-    console.log("%cHeroVau.lt/Foundry Bridge | %cusertoken: " + userToken,hvColor1,hvColor4);
-    xmlhttp.open("GET", "https://www.herovau.lt/foundrymodule.php?action=getvault&gamesystem="+encodeURIComponent(gameSystem)+"&hvVer="+hvVer+"&userToken="+encodeURIComponent(userToken), true);
+    if (hvDebug)
+      console.log("%cHeroVau.lt/Foundry Bridge | %cusertoken: " + hvUserToken,hvColor1,hvColor4);
+    xmlhttp.open("GET", heroVaultURL+"/foundrymodule.php?action=getvault&gamesystem="+encodeURIComponent(gameSystem)+"&hvVer="+hvVer+"&userToken="+encodeURIComponent(hvUserToken), true);
     xmlhttp.send();
-
 }
 
 function createPCTable(targetActor,responseJSON){
+  var charName, charRace, charClass, charLevel, charuid, pickedCharacter, selectedCharUID;
   if (hvDebug)
     console.log("%cHeroVau.lt/Foundry Bridge | %cin createPCTable",hvColor1,hvColor4);
   var htmlOut="<strong>Select a PC from the list:</strong><br><br><select name='pcid' id='pcid'>";
@@ -516,12 +686,15 @@ function createPCTable(targetActor,responseJSON){
     default: "yes",
     close: html => {
       if (pickedCharacter) {
-        console.log("yes clicked");
-        let selectedCharUID= html.find('[id="pcid"]')[0].value;
-        console.log("Selected PC id: "+selectedCharUID);
+        if (hvDebug)
+          console.log("yes clicked");
+        selectedCharUID= html.find('[id="pcid"]')[0].value;
+        if (hvDebug)
+          console.log("Selected PC id: "+selectedCharUID);
         requestCharacter(targetActor, selectedCharUID);
       } else {
-        console.log("cancel clicked");
+        if (hvDebug)
+          console.log("cancel clicked");
       }
     }
   }).render(true);
@@ -578,14 +751,15 @@ function requestCharacter(targetActor,charUID){
       // console.log("%cHeroVau.lt/Foundry Bridge | %creadyState: "+this.readyState,hvColor1,hvColor4)
       }
     };
-    console.log("%cHeroVau.lt/Foundry Bridge | %ccharUID: " + charUID,hvColor1,hvColor4)
-    xmlhttp.open("GET", "https://www.herovau.lt/foundrymodule.php?action=getCharacter&charUID="+encodeURIComponent(charUID), true);
+    if (hvDebug)
+      console.log("%cHeroVau.lt/Foundry Bridge | %ccharUID: " + charUID,hvColor1,hvColor4)
+    xmlhttp.open("GET", heroVaultURL+"/foundrymodule.php?action=getCharacter&charUID="+encodeURIComponent(charUID), true);
     xmlhttp.send();
 }
 
-
 async function importCharacter(targetActor, charURL){
-    let error=false
+    let error=false;
+    var importPCID,errMsg, charDataStr, charImport;
     var xmlhttp = new XMLHttpRequest();
     xmlhttp.onreadystatechange = function() {
       if (this.readyState == 4 && this.status == 200) {
@@ -613,22 +787,21 @@ async function importCharacter(targetActor, charURL){
         else {
           // responseJSON
           // console.log("%cHeroVau.lt/Foundry Bridge | Import ID:%c"+responseJSON._id,hvColor1,hvColor4);
-          targetPCID=targetActor.data._id;
+          let targetPCID=targetActor.data._id;
           let coreVersionMismatch=false;
           let systemVersionMismatch=false;
           let abort=false;
-          let errMsg='';
           let systemVersion=game.system.data.version;
           let coreVersion=game.data.version;
-          pcGameSystemVersion=responseJSON.flags.exportSource.systemVersion;
-          pcCoreVersion=responseJSON.flags.exportSource.coreVersion;
+          let pcGameSystemVersion=responseJSON.flags.exportSource.systemVersion;
+          let pcCoreVersion=responseJSON.flags.exportSource.coreVersion;
           if (pcCoreVersion != coreVersion){
             coreVersionMismatch=true;
             errMsg=errMsg+"This PC was exported from Foundry v"+pcCoreVersion+" and this game server is running Foundry v"+ coreVersion+".<br><br>"
           }
           if (pcGameSystemVersion != systemVersion){
             systemVersionMismatch=true;
-            if (versionCompare(pcGameSystemVersion,systemVersion) == 1) { //game system is older than PC, this could be bad!
+            if (HVversionCompare(pcGameSystemVersion,systemVersion) == 1) { //game system is older than PC, this could be bad!
               abort=true;
               errMsg=errMsg+"This PC was exported from "+game.system.data.title+": "+pcGameSystemVersion+" and this game server is running "+game.system.data.title+": "+ systemVersion+".<br><br>Unfortunately, game systems usually are not backwards compatible, so we are aborting this import. To manually override, please download the hero export from herovau.lt. <br><strong>This may break this PC -- you  have been warned!</strong><br><br>";
             } else
@@ -669,14 +842,36 @@ async function importCharacter(targetActor, charURL){
             charImport.token.sightAngle=360
           if (charImport.token.lightAngle<1)
             charImport.token.lightAngle=360
-          console.log("%cHLO Importer | %c Importing "+charImport.name,hvColor1,hvColor4);
+
+          if (hvDebug)
+            console.log("%cHLO Importer | %c Importing "+charImport.name,hvColor1,hvColor4);
           targetActor.importFromJSON(JSON.stringify(charImport));
+          var request = new XMLHttpRequest();  
+
+          request.open('GET', charImport.token.img, true);
+          request.onreadystatechange = function() {
+            if (this.status === 404) {
+              targetActor.update({'data.token.img': "icons/svg/mystery-man.svg"});
+            }
+          }
+          request.send();
+          var request2 = new XMLHttpRequest();  
+          request2.open('GET', charImport.img, true);
+          request2.onreadystatechange = function() {
+            if (this.status === 404) {
+              console.log("got a 404");
+              // console.log(targetActor.data)
+              // console.log(targetActor.img)
+              targetActor.update({'img': "icons/svg/mystery-man.svg"});
+            }
+          }
+          request2.send();
       }
         
       // console.log("%cHeroVau.lt/Foundry Bridge | %creadyState: "+this.readyState,hvColor1,hvColor4)
       }
     };
-    console.log("%cHeroVau.lt/Foundry Bridge | %cDownloading PC from: " + charURL,hvColor1,hvColor4);
+    // console.log("%cHeroVau.lt/Foundry Bridge | %cDownloading PC from: " + charURL,hvColor1,hvColor4);
     xmlhttp.open("GET", charURL, true);
     xmlhttp.send();
 }
@@ -702,7 +897,7 @@ var Cookie =
       if (host.split('.').length === 1)
       {
          // no "." in a domain - it's localhost or something similar
-         document.cookie = name+"="+value+expires+"; path=/";
+         document.cookie = name+"="+value+expires+"; path=/; SameSite=Strict";
       }
       else
       {
@@ -710,14 +905,14 @@ var Cookie =
          domainParts.shift();
          domain = '.'+domainParts.join('.');
 
-         document.cookie = name+"="+value+expires+"; path=/; domain="+domain;
+         document.cookie = name+"="+value+expires+"; path=/; domain="+domain+"; SameSite=Strict";
          // check if cookie was successfuly set to the given domain
          // (otherwise it was a Top-Level Domain)
          if (Cookie.get(name) == null || Cookie.get(name) != value)
          {
             // append "." to current domain
             domain = '.'+host;
-            document.cookie = name+"="+value+expires+"; path=/; domain="+domain;
+            document.cookie = name+"="+value+expires+"; path=/; domain="+domain+"; SameSite=Strict";
          }
       }
    },
@@ -749,10 +944,10 @@ var Cookie =
  * Simply compares two string version values.
  * 
  * Example:
- * versionCompare('1.1', '1.2') => -1
- * versionCompare('1.1', '1.1') =>  0
- * versionCompare('1.2', '1.1') =>  1
- * versionCompare('2.23.3', '2.22.3') => 1
+ * HVversionCompare('1.1', '1.2') => -1
+ * HVversionCompare('1.1', '1.1') =>  0
+ * HVversionCompare('1.2', '1.1') =>  1
+ * HVversionCompare('2.23.3', '2.22.3') => 1
  * 
  * Returns:
  * -1 = left is LOWER than right
@@ -767,7 +962,7 @@ var Cookie =
  * @author Alexey Bass (albass)
  * @since 2011-07-14
  */
-versionCompare = function(left, right) {
+let HVversionCompare = function(left, right) {
     if (typeof left + typeof right != 'stringstring')
         return false;
     
@@ -785,3 +980,55 @@ versionCompare = function(left, right) {
     
     return 0;
 }
+
+function getHash(encodedHeroJSON) {
+  var xmlhttp = new XMLHttpRequest();
+  xmlhttp.onreadystatechange = function() {
+  if (this.readyState == 4 && this.status == 200) {
+    let responseJSON = JSON.parse(this.responseText);
+    if (hvDebug) 
+      console.log("%cHeroVau.lt/Foundry Bridge | %c"+JSON.stringify(responseJSON),hvColor1,hvColor4);
+    let newHash=responseJSON.newHash;  
+    return newHash;
+    }
+  };
+  xmlhttp.open("POST", heroVaultURL+"/foundrymodule.php", true);
+  xmlhttp.setRequestHeader("Content-type", "application/x-www-form-urlencoded");
+  xmlhttp.send('action=genHash&encodedChar='+encodedHeroJSON);
+}
+
+export function exportToHVFromPBHLO(heroJSON,tAct) {
+  let error=false;
+  let action='importNewPCFromPBHLO';
+  var xmlhttp = new XMLHttpRequest();
+  const gameSystem=game.data.system.id;
+
+  let pcEncodedJSON=encodeURIComponent(heroJSON);
+  let newHash=getHash(pcEncodedJSON);
+  // console.log(pcEncodedJSON);
+  xmlhttp.onreadystatechange = function() {
+    if (this.readyState == 4 && this.status == 200) {
+      let responseJSON = JSON.parse(this.responseText);
+      if (responseJSON.error==false){
+        ui.notifications.info("Successfully exported to HeroVau.lt."); 
+        tAct.update({
+          'flags.herovault.uid':responseJSON.charhash
+        });
+        return responseJSON.charhash;
+      } else {
+        ui.notifications.warn("Unable to export to HeroVau.lt. Please try manually."); 
+      }
+    }
+  };
+  // console.log("%cHeroVau.lt/Foundry Bridge | %chttps://herovau.lt/foundrymodule.php?action=importNewPC&userToken="+userToken+"&encodedChar="+pcEncodedJSON,hvColor1,hvColor4);
+  xmlhttp.open("POST", heroVaultURL+"/foundrymodule.php", true);
+  xmlhttp.setRequestHeader("Content-type", "application/x-www-form-urlencoded");
+  xmlhttp.send('action='+action+'&userToken='+hvUserToken+'&encodedChar='+pcEncodedJSON+'&gamesystem='+gameSystem+"&charUID="+newHash);
+  
+}
+
+export function supportCheck() {
+  // console.log("SupportCheck: "+hvUserToken)
+  if(hvUserToken) {
+    return true;
+  }